moving out all diff attn code to plugin repo
